refactor: extract masked email state hook
