+++ src/pages/popup/components/home/top/Top.tsx
@@ -1,25 +1,13 @@
 import { useState } from 'react';
 import SearchBar from '@pages/popup/components/home/top/SearchBar';
 import SettingsDropdown from '@pages/popup/components/home/top/SettingsDropdown';
-import { MaskedEmail } from 'fastmail-masked-email';
 import icon from '@assets/img/icon.svg';
 
 interface Props {
   onSearchChange: (searchQuery: string) => void;
   onRefresh: () => Promise<void>;
-  onLogout: () => void;
-  addNewEmailToEmailList: (newEmail: MaskedEmail) => void;
-  setSelectedEmail: (
-    value:
-      | ((prevState: MaskedEmail | null) => MaskedEmail | null)
-      | MaskedEmail
-      | null
-  ) => void;
-  activeTabUrl: string;
   openLogoutConfirmationModal: () => void;
-  closeLogoutConfirmationModal: () => void;
   openCreateEmailModal: () => void;
-  closeCreateEmailModal: () => void;
 }
 
 export default function TopComponent({
